+++ src/roslibpy/comm/comm.py
@@ -3,7 +3,14 @@ from __future__ import print_function
 import json
 import logging
 
-from roslibpy.core import Message, MessageEncoder, ServiceResponse
+from roslibpy.core import (
+    ActionFeedback,
+    ActionGoalStatus,
+    ActionResult,
+    Message,
+    MessageEncoder,
+    ServiceResponse,
+)
 
 LOGGER = logging.getLogger("roslibpy")
 
@@ -22,19 +29,23 @@ class RosBridgeProtocol(object):
         super(RosBridgeProtocol, self).__init__(*args, **kwargs)
         self.factory = None
         self._pending_service_requests = {}
+        self._pending_action_requests = {}
         self._message_handlers = {
             "publish": self._handle_publish,
             "service_response": self._handle_service_response,
             "call_service": self._handle_service_request,
+            "send_action_goal": self._handle_action_request,  # TODO: action server
+            "cancel_action_goal": self._handle_action_cancel,  # TODO: action server
+            "action_feedback": self._handle_action_feedback,
+            "action_result": self._handle_action_result,
+            "status": None,  # TODO: add handlers for op: status
         }
-        # TODO: add handlers for op: status
 
     def on_message(self, payload):
         message = Message(json.loads(payload.decode("utf8")))
         handler = self._message_handlers.get(message["op"], None)
         if not handler:
             raise RosBridgeException('No handler registered for operation "%s"' % message["op"])
-
         handler(message)
 
     def send_ros_message(self, message):
@@ -106,3 +117,59 @@ class RosBridgeProtocol(object):
             raise ValueError("Expected service name missing in service request")
 
         self.factory.emit(message["service"], message)
+
+    def send_ros_action_goal(self, message, resultback, feedback, errback):
+        """Initiate a ROS action request by sending a goal through the ROS Bridge.
+
+        Args:
+        message (:class:`.Message`): ROS Bridge Message containing the action request.
+        callback: Callback invoked on receiving result.
+        feedback: Callback invoked when receiving feedback from action server.
+        errback: Callback invoked on error.
+        """
+        request_id = message["id"]
+        self._pending_action_requests[request_id] = (resultback, feedback, errback)
+
+        json_message = json.dumps(dict(message), cls=MessageEncoder).encode("utf8")
+        LOGGER.debug("Sending ROS action goal request: %s", json_message)
+
+        self.send_message(json_message)
+
+    def _handle_action_request(self, message):
+        if "action" not in message:
+            raise ValueError("Expected action name missing in action request")
+        raise RosBridgeException('Action server capabilities not yet implemented')
+
+    def _handle_action_cancel(self, message):
+        if "action" not in message:
+            raise ValueError("Expected action name missing in action request")
+        raise RosBridgeException('Action server capabilities not yet implemented')
+
+    def _handle_action_feedback(self, message):
+        if "action" not in message:
+            raise ValueError("Expected action name missing in action feedback")
+
+        request_id = message["id"]
+        _, feedback, _ = self._pending_action_requests.get(request_id, None)
+        feedback(ActionFeedback(message["values"]))
+
+    def _handle_action_result(self, message):
+        request_id = message["id"]
+        action_handlers = self._pending_action_requests.get(request_id, None)
+
+        if not action_handlers:
+            raise RosBridgeException('No handler registered for action request ID: "%s"' % request_id)
+
+        resultback, _ , errback = action_handlers
+        del self._pending_action_requests[request_id]
+
+        LOGGER.debug("Received Action result with status: %s", message["status"])
+
+        results = {"status": ActionGoalStatus(message["status"]).name, "values": message["values"]}
+
+        if "result" in message and message["result"] is False:
+            if errback:
+                errback(results)
+        else:
+            if resultback:
+                resultback(ActionResult(results))